Refactor booking models: move model classes to BookingModels.cs and remove redundant definitions from BookingsController.cs
