+++ Models/BookingModels.cs
@@ -1,0 +1,59 @@
+﻿using MongoDB.Bson;
+using MongoDB.Bson.Serialization.Attributes;
+using System;
+
+namespace SparkPoint_Server.Models
+{
+    public class Booking
+    {
+        [BsonId]
+        [BsonRepresentation(BsonType.ObjectId)]
+        public string Id { get; set; }
+
+        [BsonElement("ownerNIC")]
+        public string OwnerNIC { get; set; }
+
+        [BsonElement("stationId")]
+        [BsonRepresentation(BsonType.ObjectId)]
+        public string StationId { get; set; }
+
+        [BsonElement("reservationTime")]
+        public DateTime ReservationTime { get; set; }
+
+        [BsonElement("status")]
+        public string Status { get; set; }
+
+        [BsonElement("createdAt")]
+        [BsonDateTimeOptions(Kind = DateTimeKind.Utc)]
+        public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
+
+        [BsonElement("updatedAt")]
+        [BsonDateTimeOptions(Kind = DateTimeKind.Utc)]
+        public DateTime UpdatedAt { get; set; } = DateTime.UtcNow;
+    }
+    public class BookingCreateModel
+    {
+        public string OwnerNIC { get; set; }
+        public string StationId { get; set; }
+        public DateTime ReservationTime { get; set; }
+    }
+
+    public class BookingUpdateModel
+    {
+        public string StationId { get; set; }
+        public DateTime? ReservationTime { get; set; }
+    }
+
+    public class BookingStatusUpdateModel
+    {
+        public string Status { get; set; }
+    }
+
+    public class BookingFilterModel
+    {
+        public string Status { get; set; }
+        public string StationId { get; set; }
+        public DateTime? FromDate { get; set; }
+        public DateTime? ToDate { get; set; }
+    }
+}
